+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Translator can now put emphasis on whole phrases.
 - Fix "(word) a" not being parsed properly.
 - Proper name now can't modify adjective translations &ndash; translations like "good named Nimi" are now removed.
+- Untranslatable phrases now handled as untranslatable error
+- New kind of error where the sentence can't be translated but it should be
+- Minor fixes and changes
 
 ## 0.0.1